pandemic_api/requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
# GDAL==3.2.0
# GeoIP==1.3.2
# geoip2==4.1.0
idna==2.10
maxminddb==2.0.3
multidict==5.1.0
protobuf==3.14.0
pycodestyle==2.6.0
pymongo==3.11.2
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
